requirements.txt
aiogram
requests
telethon
python-dotenv
